Reverse refactor
