+++ packages/excalidraw/components/Stats/Angle.tsx
@@ -11,13 +11,8 @@ import type Scene from "@excalidraw/element/Scene";
 
 import { angleIcon } from "../icons";
 
-import DragInput from "./DragInput";
-import {
-  type DragInputCallbackType,
-  getStepSizedValue,
-  isPropertyEditable,
-  updateBindings,
-} from "./utils";
+import DragInput, { type DragInputCallbackType } from "./DragInput";
+import { getStepSizedValue, isPropertyEditable, updateBindings } from "./utils";
 
 import type { AppState } from "../../types";
