+++ packages/excalidraw/components/Stats/Dimension.tsx
@@ -12,12 +12,8 @@ import type { ExcalidrawElement } from "@excalidraw/element/types";
 
 import type Scene from "@excalidraw/element/Scene";
 
-import DragInput from "./DragInput";
-import {
-  type DragInputCallbackType,
-  getStepSizedValue,
-  isPropertyEditable,
-} from "./utils";
+import DragInput, { type DragInputCallbackType } from "./DragInput";
+import { getStepSizedValue, isPropertyEditable } from "./utils";
 
 import type { AppState } from "../../types";
