+++ packages/excalidraw/components/Stats/FontSize.tsx
@@ -16,8 +16,8 @@ import type Scene from "@excalidraw/element/Scene";
 
 import { fontSizeIcon } from "../icons";
 
-import StatsDragInput from "./DragInput";
-import { type DragInputCallbackType, getStepSizedValue } from "./utils";
+import StatsDragInput, { type DragInputCallbackType } from "./DragInput";
+import { getStepSizedValue } from "./utils";
 
 import type { AppState } from "../../types";
