+++ packages/excalidraw/components/Stats/MultiDimension.tsx
@@ -24,9 +24,9 @@ import type {
 
 import type Scene from "@excalidraw/element/Scene";
 
-import DragInput from "./DragInput";
+import DragInput, { type DragInputCallbackType } from "./DragInput";
 import { getAtomicUnits, getStepSizedValue, isPropertyEditable } from "./utils";
-import { type DragInputCallbackType, getElementsInAtomicUnit } from "./utils";
+import { getElementsInAtomicUnit } from "./utils";
 
 import type { AtomicUnit } from "./utils";
 import type { AppState } from "../../types";
